+++ dephell/commands/autocomplete.py
@@ -1,6 +1,7 @@
 # built-in
 from argparse import ArgumentParser
 from pathlib import Path
+from platform import platform
 
 # external
 from appdirs import user_data_dir
@@ -47,10 +48,16 @@ class AutocompleteCommand(BaseCommand):
 
     def _bash(self):
         script = make_bash_autocomplete()
+
+        # https://github.com/dephell/dephell/pull/62
         path = Path.home() / '.local' / 'etc' / 'bash_completion.d' / 'dephell.bash-completion'
+        if platform().lower() == 'darwin':
+            # ref. https://itnext.io/programmable-completion-for-bash-on-macos-f81a0103080b
+            path = Path('/') / 'usr' / 'local' / 'etc' / 'bash_completion.d' / 'dephell.bash-completion'
+        path.parent.mkdir(parents=True, exist_ok=True)
         path.write_text(script)
 
-        for rc_name in ('.bashrc', '.profile'):
+        for rc_name in ('.bashrc', '.profile', '.bash_profile'):
             rc_path = Path.home() / rc_name
             if not rc_path.exists():
                 continue