+++ tests/test_repositories/test_local.py
@@ -10,7 +10,7 @@ loop = asyncio.get_event_loop()
 
 
 def test_root_file():
-    repo = LocalRepo(Path() / 'tests' / 'requirements' / 'setup.py')
+    repo = LocalRepo(Path('tests') / 'requirements' / 'setup.py')
     root = repo.get_root(name='dephell', version='0.2.0')
     assert root.name == 'dephell'
     assert root.version == '0.2.0'
@@ -18,7 +18,7 @@ def test_root_file():
 
 
 def test_root_dir():
-    repo = LocalRepo(Path() / 'tests' / 'requirements' / 'egg-info')
+    repo = LocalRepo(Path('tests') / 'requirements' / 'egg-info')
     root = repo.get_root(name='dephell', version='0.2.0')
     assert root.name == 'dephell'
     assert root.version == '0.2.1'
@@ -26,8 +26,9 @@ def test_root_dir():
 
 
 def test_deps_file():
-    repo = LocalRepo(Path() / 'tests' / 'requirements' / 'setup.py')
+    repo = LocalRepo(Path('tests') / 'requirements' / 'setup.py')
     coroutine = repo.get_dependencies(name='dephell', version='0.2.0')
     deps = loop.run_until_complete(asyncio.gather(coroutine))[0]
     deps = {dep.name: dep for dep in deps}
-    assert set(deps) == {'attrs', 'cached-property', 'packaging', 'requests', 'colorama'}
+    expected = {'attrs', 'cached-property', 'packaging', 'requests', 'colorama', 'libtest'}
+    assert set(deps) == expected